requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
